Warning fixes.

+++ include/inviwo/qt/widgets/properties/ordinalminmaxpropertywidgetqt.h
@@ -181,7 +181,7 @@ protected:
         if (periodPosition == std::string::npos) 
             return 0;
         else
-            return str.length() - periodPosition - 1;
+            return static_cast<int>(str.length() - periodPosition - 1);
     }
     virtual std::string getToolTipText();
